break auth components into seperate files, have template for sign up
