+++ public/js/app.js
@@ -8,6 +8,9 @@ const Route = ReactRouter.Route;
 const Link = ReactRouter.Link;
 const auth = require('./auth_helpers');
 const $ = require('jquery');
+const Login = require('./login.js');
+const Logout = require('./logout.js');
+const SignUp = require('./signup.js');
 
 const App = React.createClass({
   getInitialState: function() {
@@ -38,6 +41,7 @@ const App = React.createClass({
               <Link to="/login">Sign in</Link>
             )}
           </li>
+          <li><Link to="/new">Sign Up</Link></li>
           <li><Link to="/dashboard">Dashboard</Link> (authenticated)</li>
         </ul>
         {this.props.children || <p>You are {!this.state.loggedIn && 'not'} logged in.</p>}
@@ -81,62 +85,6 @@ const Dashboard = React.createClass({
   }
 })
 
-const Login = React.createClass({
-
-  contextTypes: {
-    router: React.PropTypes.object.isRequired
-  },
-
-  getInitialState: function() {
-    return {
-      error: false
-    }
-  },
-
-  handleSubmit: function(event) {
-    event.preventDefault()
-
-    const email = this.refs.email.value
-    const pass = this.refs.pass.value
-
-    auth.login(email, pass, (loggedIn) => {
-      if (!loggedIn)
-        return this.setState({ error: true })
-
-      const { location } = this.props
-
-      if (location.state && location.state.nextPathname) {
-        this.context.router.replace(location.state.nextPathname)
-      } else {
-        this.context.router.replace('/')
-      }
-    })
-  },
-
-  render: function() {
-    return (
-      <form onSubmit={this.handleSubmit}>
-        <label><input ref="email" placeholder="email" defaultValue="test2" /></label>
-        <label><input ref="pass" placeholder="password" /></label> (hint: password1)<br />
-        <button type="submit">login</button>
-        {this.state.error && (
-          <p>Bad login information</p>
-        )}
-      </form>
-    )
-  }
-})
-
-
-const Logout = React.createClass({
-  componentDidMount: function() {
-    auth.logout()
-  },
-
-  render: function() {
-    return <p>You are now logged out</p>
-  }
-})
 
 function requireAuth(nextState, replace) {
   if (!auth.loggedIn()) {
@@ -152,6 +100,7 @@ ReactDOM.render((
     <Route path="/" component={App}>
       <Route path="login" component={Login} />
       <Route path="logout" component={Logout} />
+      <Route path="new" component={SignUp} />
       <Route path="dashboard" component={Dashboard} onEnter={requireAuth} />
     </Route>
   </Router>
